Use the first token to find child, avoiding traversing all children.