@merrymercy 2024-04-16
@ispobock This seems to be an interesting optimization.
Did you check the correctness of many workloads? Did you notice any performance improvements?