@ispobock 2024-04-17
@merrymercy I tested on mmlu dataset with different parallel, the accuracy is the same with the main branch. The total latency has slight improvement.

| with optimization | parallel | total latency | avg accuracy  |
|-------------------|----------|---------------|---------------|
| N                 | 64       | 258.204       | 0.528         |
| Y                 | 64       | 257.729       | 0.528         |
| N                 | 128      | 254.929       | 0.528         |
| Y                 | 128      | 254.867       | 0.528         |

I also test on sharedgpt dataset with 1000 prompts and found the throughput has slight improvement.
without optimization:
```
Total time: 175.29 s
Throughput: 5.70 requests/s
Average latency: 65.24 s
Average latency per token: 0.18 s
Average latency per output token: 1.10 s
```
with optimization:
```
Total time: 173.21 s
Throughput: 5.77 requests/s
Average latency: 64.15 s
Average latency per token: 0.17 s
Average latency per output token: 1.08 s
```

Test env:
- model: llama2_13b_chat
- gup: 1 A100 80G

Reproduce:
```
python -m sglang.launch_server --model-path /workdir/llama2_13b_chat/ --port 30000

python benchmark/mmlu/bench_sglang.py

python benchmark/latency_throughput/bench_throughput.py --backend srt --tokenizer /workdir/llama2_13b_chat/ --dataset /workdir/ShareGPT_V3_unfiltered_cleaned_split.json --num-prompts 1000 --request-rate 128 --port 30000
```