@ispobock 2024-04-17
For 5000 prompts of sharegpt dataset benchmark, the throughput is improved **11.2%** with this optimization.
without optimization:
```
Total time: 973.67 s
Throughput: 5.14 requests/s
Average latency: 439.68 s
Average latency per token: 1.17 s
Average latency per output token: 10.23 s
```
with optimization:
```
Total time: 874.84 s
Throughput: 5.72 requests/s
Average latency: 382.35 s
Average latency per token: 1.01 s
Average latency per output token: 8.91 s
```